backend/todo/requirements.txt
black==19.10b0
django==3.0.8
django-cors-headers==3.4.0
django-extensions==2.2.9
djangorestframework==3.11.0
factory-boy==2.12.0
